Add function validation to get-manifest hook. Move function validation tests to utilities tests.
