+++ src/get_manifest.ts
@@ -1,5 +1,8 @@
 import { deepMerge, getProtocolInterface, path } from "./deps.ts";
-import { getDefaultExport } from "./utilities.ts";
+import {
+  forEachValidatedManifestFunction,
+  getDefaultExport,
+} from "./utilities.ts";
 
 // Responsible for taking a working directory, and an output directory
 // and placing a manifest.json in the root of the output directory
@@ -121,7 +124,8 @@ async function readImportedManifestFile(filename: string) {
 if (import.meta.main) {
   const protocol = getProtocolInterface(Deno.args);
   const generatedManifest = await getManifest(Deno.cwd());
-  // TODO: validate function paths
+  // validate functions referenced in the manifest
+  forEachValidatedManifestFunction(Deno.cwd(), generatedManifest);
   const prunedManifest = cleanManifest(generatedManifest);
   protocol.respond(JSON.stringify(prunedManifest));
 }
