+++ src/tests/build_test.ts
@@ -1,7 +1,6 @@
 import { validateAndCreateFunctions } from "../build.ts";
 import {
   assertExists,
-  assertRejects,
   assertSpyCalls,
   returnsNext,
   stub,
@@ -56,192 +55,5 @@ Deno.test("build hook tests", async (t) => {
         runStub.restore();
       },
     );
-
-    await tt.step(
-      "should throw an exception when fed a function file that does not have a default export",
-      async () => {
-        const protocol = MockProtocol();
-        const manifest = {
-          "functions": {
-            "test_function": {
-              "title": "Test function",
-              "description": "this is a test",
-              "source_file":
-                "src/tests/fixtures/functions/test_function_no_export_file.ts",
-              "input_parameters": {
-                "required": [],
-                "properties": {},
-              },
-              "output_parameters": {
-                "required": [],
-                "properties": {},
-              },
-            },
-          },
-        };
-        const outputDir = await Deno.makeTempDir();
-        await assertRejects(
-          () =>
-            validateAndCreateFunctions(
-              Deno.cwd(),
-              outputDir,
-              manifest,
-              protocol,
-            ),
-          Error,
-          "no default export",
-        );
-      },
-    );
-
-    await tt.step(
-      "should throw an exception when fed a function file that has a non-function default export",
-      async () => {
-        const protocol = MockProtocol();
-        const manifest = {
-          "functions": {
-            "test_function": {
-              "title": "Test function",
-              "description": "this is a test",
-              "source_file":
-                "src/tests/fixtures/functions/test_function_not_function_file.ts",
-              "input_parameters": {
-                "required": [],
-                "properties": {},
-              },
-              "output_parameters": {
-                "required": [],
-                "properties": {},
-              },
-            },
-          },
-        };
-        const outputDir = await Deno.makeTempDir();
-        await assertRejects(
-          () =>
-            validateAndCreateFunctions(
-              Deno.cwd(),
-              outputDir,
-              manifest,
-              protocol,
-            ),
-          Error,
-          "default export is not a function",
-        );
-      },
-    );
-
-    await tt.step(
-      "should throw an exception when manifest entry for a function points to a non-existent file",
-      async () => {
-        const protocol = MockProtocol();
-        const manifest = {
-          "functions": {
-            "test_function": {
-              "title": "Test function",
-              "description": "this is a test",
-              "source_file":
-                "src/tests/fixtures/functions/test_function_this_file_does_not_exist.ts",
-              "input_parameters": {
-                "required": [],
-                "properties": {},
-              },
-              "output_parameters": {
-                "required": [],
-                "properties": {},
-              },
-            },
-          },
-        };
-        const outputDir = await Deno.makeTempDir();
-        await assertRejects(
-          () =>
-            validateAndCreateFunctions(
-              Deno.cwd(),
-              outputDir,
-              manifest,
-              protocol,
-            ),
-          Error,
-          "Could not find file",
-        );
-      },
-    );
-
-    await tt.step(
-      "should throw an exception when manifest entry for a function does not have a source_file defined",
-      async () => {
-        const protocol = MockProtocol();
-        const manifest = {
-          "functions": {
-            "test_function": {
-              "title": "Test function",
-              "description": "this is a test",
-              "input_parameters": {
-                "required": [],
-                "properties": {},
-              },
-              "output_parameters": {
-                "required": [],
-                "properties": {},
-              },
-            },
-          },
-        };
-        const outputDir = await Deno.makeTempDir();
-        await assertRejects(
-          () =>
-            validateAndCreateFunctions(
-              Deno.cwd(),
-              outputDir,
-              manifest,
-              protocol,
-            ),
-          Error,
-          "No source_file property provided",
-        );
-      },
-    );
-
-    await tt.step("should not write out functions of type 'API'", async () => {
-      const protocol = MockProtocol();
-      const manifest = {
-        "functions": {
-          "test_function": {
-            "title": "Test function",
-            "description": "this is a test",
-            "source_file":
-              "src/tests/fixtures/functions/test_function_not_function_file.ts",
-            "type": "API",
-            "input_parameters": {
-              "required": [],
-              "properties": {},
-            },
-            "output_parameters": {
-              "required": [],
-              "properties": {},
-            },
-          },
-        },
-      };
-      const outputDir = await Deno.makeTempDir();
-      await validateAndCreateFunctions(
-        Deno.cwd(),
-        outputDir,
-        manifest,
-        protocol,
-      );
-      // Stub out call to `Deno.run` and if invoked, throw an error which should fail the test
-      const runStub = stub(
-        Deno,
-        "run",
-        returnsNext([
-          new Error(
-            "Deno.run should not be invoked, test should fail",
-          ) as unknown as Deno.Process<Deno.RunOptions>,
-        ]),
-      );
-      assertSpyCalls(runStub, 0);
-    });
   });
 });
